refactor: 테스트 mock data 리팩토링
- 테스트 케이스별로 테스트 mock data 확인할 수 있도록 수정
