+++ __tests__/cartItemDelete.test.tsx
@@ -5,9 +5,8 @@ import { act } from '@testing-library/react';
 import { useRecoilValue } from 'recoil';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { INITIAL_ITEMS } from './mockData/cartItems';
-import { localStorageMock } from './utils/localStorageMock';
-import { renderHookWithRecoilRoot } from './utils/recoilTestUtils';
+import { INITIAL_ITEMS } from './mockData';
+import { localStorageMock, renderHookWithRecoilRoot } from './utils';
 
 const renderUseDeleteCartItem = (id: number) => {
   const { result } = renderHookWithRecoilRoot(() => {
