+++ __tests__/mockData/index.ts
@@ -1,0 +1,2 @@
+export * from './cartItems';
+export * from './coupon';
